lsp: Port existing unit tests for reStructuredText regexes
